pages controller draft, exception handling

+++ src/main/java/ru/venkispb/server/repository/AdminRepository.java
@@ -3,5 +3,5 @@ package ru.venkispb.server.repository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import ru.venkispb.server.entity.AdminsEntity;
 
-public interface AdminRepository extends PagingAndSortingRepository<AdminsEntity, Long> {
+public interface AdminRepository extends PagingAndSortingRepository<AdminsEntity, Integer> {
 }
